backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlmodel==0.0.24
python-multipart==0.0.20
httpx==0.28.1
openai==1.82.0
rembg>=2.0.68
onnxruntime>=1.21.1
Pillow==11.2.1
aiofiles==24.1.0
